Move fluent custom resources related component and files into dedicated package
