+++ pkg/component/observability/logging/fluentcustomresources/fluentcustomresources_suite_test.go
@@ -1,4 +1,4 @@
-// Copyright 2023 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
+// Copyright 2024 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package customresources_test
+package fluentcustomresources_test
 
 import (
 	"testing"
@@ -21,7 +21,7 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-func TestCustomResources(t *testing.T) {
+func TestFluentCustomresources(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "Component Logging FluentOperator CustomResources Suite")
+	RunSpecs(t, "Component Observability Logging FluentCustomResources Suite")
 }
