+++ skaffold.yaml
@@ -917,7 +917,7 @@ build:
             - pkg/component/observability/logging/eventlogger
             - pkg/component/observability/logging/fluentbit
             - pkg/component/observability/logging/fluentoperator
-            - pkg/component/observability/logging/fluentoperator/customresources
+            - pkg/component/observability/logging/fluentcustomresources
             - pkg/component/observability/logging/vali
             - pkg/component/observability/logging/vali/constants
             - pkg/component/observability/monitoring
